test(rsc): set up co-located client HMR regression via setupInlineFixture
