--- packages/plugin-rsc/examples/co-located-client-hmr/.gitignore
@@ -1,2 +1,0 @@
-node_modules
-dist
