--- packages/plugin-rsc/examples/co-located-client-hmr/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@vitejs/plugin-rsc-examples-co-located-client-hmr",
-  "version": "0.0.0",
-  "private": true,
-  "license": "MIT",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "react": "^19.2.7",
-    "react-dom": "^19.2.7"
-  },
-  "devDependencies": {
-    "@types/react": "^19.2.17",
-    "@types/react-dom": "^19.2.3",
-    "@vitejs/plugin-react": "latest",
-    "@vitejs/plugin-rsc": "latest",
-    "rsc-html-stream": "^0.0.7",
-    "vite": "^8.0.16"
-  }
-}
