--- packages/plugin-rsc/examples/co-located-client-hmr/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "compilerOptions": {
-    "erasableSyntaxOnly": true,
-    "allowImportingTsExtensions": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "skipLibCheck": true,
-    "verbatimModuleSyntax": true,
-    "noEmit": true,
-    "moduleResolution": "Bundler",
-    "module": "ESNext",
-    "target": "ESNext",
-    "lib": ["ESNext", "DOM"],
-    "types": ["vite/client", "@vitejs/plugin-rsc/types"],
-    "jsx": "react-jsx"
-  }
-}
